MypageとBlogListのリファクタリング
